refactor(tui): extract editor history boundary
